pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "baf-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pyyaml",
  "requests",
]